/ppbox/branches/1.2.0 /src/ppbox/demux 解封装模块整理
git-svn-id: http://svn01/repositories/r01/ppbox/branches/1.2.0@5063 b3a1389e-2159-5c43-9038-463bf877d0d9

+++ base/BytesStream.h
@@ -55,7 +55,7 @@ namespace ppbox
                 pos_type pos)
             {
                 boost::uint64_t pos64 = pos;
-                boost::uint32_t off32 = 0;
+                boost::uint64_t off32 = 0;
                 boost::asio::const_buffer buf;;
                 boost::system::error_code ec;
                 ec = buffer_.segment_buffer(segment_, type, pos64, off32, buf);
